Constructors

SplineGrids.KnotVector — Method

KnotVector(n_basis_functions::Integer, degree::Integer; extent::Tuple{Number, Number} = (0,1), distribution::Symbol = :equispaced)

Construct a clamped knot vector, i.e. the multiplicity of the first and last knot is degree + 1 and the other multiplicities are 1.

Arguments

n_basis_functions: The number of basis functions that will be defined on this knot vector

degree: The degree of the basis functions that will be defined on this knot vector

Keyword Arguments

extent: A tuple (tmin, tmax) defining the extend of the knot vector

distribution: The distribution of the internal knots. The options are :equispaced or :random

SplineGrids.SplineDimension — Method

SplineDimension(n_basis_functions::Integer, degree::Integer, n_sample_points::Integer; kwargs...)::SplineDimension

Constructor for a SplineDimension. For now the sample points are evenly spaced on the extent of the knot vector. Key word arguments are passed to the KnotVector constructor.

SplineGrids.SplineGrid — Method

SplineGrid(spline_dimensions::NTuple{N_in, <:SplineDimension}, dim_out::Integer)

Define a SplineGrid from an NTuple of spline dimensions and the number of output dimensions.

Inputs

spline_dimensions: an NTuple of spline dimensions

Nout: The number of output dimensions. I.e. the control points and thus the spline live in ℝ^Nout.

Evaluation

SplineGrids.evaluate! — Method

evaluate!(spline_dimension)

Per sample point, get the value of the spline_dimension.degree + 1 basis functions that have a non-zero value for that sample point. This is based on the Cox-de Boor recursion formula.

The l-th sample point t has sample index i, meaning that t ∈ [tᵢ, tᵢ₊₁). Therefore Bᵢ₀(t) = 1, Bⱼ₀(t) = 0 for j ≠ i. For degree k, t is in the domain of Bⱼₖ which is [tⱼ, tⱼ₊ₖ₊₁), for j = i - k, ..., i.

Arguments

spline_dimension

SplineGrids.evaluate! — Method

evaluate!spline_grid::AbstractSplineGrid{Nin};
    derivative_order::NTuple{Nin, <:Integer} = ntuple(_ -> 0, Nin),
    control_points::AbstractArray = spline_grid.control_points,
    eval::AbstractArray = spline_grid.eval)

Evaluate the spline grid, that is: take the evaluated basis functions for each sample point for each SplineDimension, and compute the output grid on each sample point combination as a linear combination of control with basis function products as coefficients.

Uses the control_points and eval arrays from the spline_grid by default, but different arrays can be specified as a convencience for optimization algorithms.

SplineGrids.evaluate_adjoint! — Method

evaluate_adjoint!(spline_grid::AbstractSplineGrid{Nin};
    derivative_order::NTuple{Nin, <:Integer} = ntuple(_ -> 0, Nin),
    control_points::AbstractArray = spline_grid.control_points,
    eval::AbstractArray = spline_grid.eval)

evaluate the adjoint of the linear mapping control_points -> eval. This is a computation of the form eval -> control_points. If we write evaluate!(spline_grid) as a matrix vector multiplication eval = M * control_points, Then the adjoint is given by v -> M' * v. This mapping is used in fitting algorithms.

Structs

SplineGrids.KnotVector — Type

KnotVector(knots, multiplicities)

Defines a knot vector.

Arguments

knot_values: The values in the knot vector. Must be strictly increasing.

multiplicities: The multiplicity of each knot in knots.

SplineGrids.SplineDimension — Type

SplineDimension(degree, knot_vector, sample_points, sample_indices)

Defines the set of basis functions for a single dimension, and how it is sampled.

Arguments

degree: The degree of the piecewise polynomial basis functions.

max_derivative_order: The maximum derivative order of the basis functions that will be computed.

knot_vector: The knot vector on which the basis functions are defined.

sample_points: The points in the domain of the basis functions where they are sampled. Must

lie within the boundaries of the knot vector.

sample_indices: The indices i of the sample points t in the knot vector such that knot_vector.knots[i] ≤ t < knot_vector.knots[i + 1]`

eval: An array of shape (length(sample_points), degree + 1, max_derivative + 1), with per sample point the values of those basis functions

eval_prev: Helper array for intermediate results in the basis function computations whose support the sample point is in, and the derivatives if requested.

SplineGrids.SplineGrid — Type

The SplineGrid is the central object of the SplineGrids.jl package, containing all information to evaluate the defined spline on the defined grid.

Fields

spline_dimensions: A SplineDimension per dimension of the spline, containing data to evaluate basis functions.

sample_indices: For each global sample point, the linear index in the control_points array before it is offset for a particular index in the control point kernel and output dimension.

control points: The points that define the shape of the spline, and in how many dimensions it is embedded.

weights: For now unsupported, will eventually be used to define NURBS.

eval: The array where the evaluated spline grid is stored.

basis_function_products: An array of intermediate results for evaluating the spline grids, containing products of basis functions from the various spline dimensions.

Utility functions

SplineGrids.decompress — Function

Transform spline_dimension.eval into a matrix of shape (n_sample_points, n_points - degree - 1) which explicitly gives the value for each basis function at each sample point.